requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Google Sheets integration (optional)
google-auth>=2.22.0
google-api-python-client>=2.100.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
